+++ src/s11/s11_common.c
@@ -38,7 +38,7 @@
 nw_rc_t
 s11_ie_indication_generic (
   uint8_t ieType,
-  uint8_t ieLength,
+  uint16_t ieLength,
   uint8_t ieInstance,
   uint8_t * ieValue,
   void *arg)
